+++ src/components/PropertyWindow.tsx
@@ -1,7 +1,6 @@
 import React from 'react';
-import { Property } from "../lib/core/Property";
+import { Property, FabricContext } from "../lib/core";
 import { ExposedProperty } from './ExposedProperty';
-import { FabricContext } from '../lib/core/FabricContext';
 
 export type PropertyWindowsProps = {
     context: FabricContext,
