+++ src/lib/core/FabricRelated/EditorObject.ts
