+++ src/lib/core/FabricRelated/Plugin.ts
