+++ src/lib/core/FabricRelated/index.ts
@@ -1,0 +1,5 @@
+export * from './EditorObject';
+export * from './FabricContextUser';
+export * from './IDestroyable';
+export * from './Plugin';
+export * from './Property';
