+++ src/lib/core/index.ts
@@ -1,8 +1,4 @@
-export * from './Plugin';
-export * from './Property';
-export * from './EditorObject';
-export * from './FabricContext';
+export * from './FabricRelated';
+export * from './FabricRelated/FabricContext';
 export * from './commands';
 export * from './persistence';
-export * from './FabricContextUser';
-export * from './IDestroyable';
